+++ CHANGELOG.md
@@ -6,6 +6,7 @@ By far most changes relate to `atomic-server`, so if not specified, assume the c
 ## v0.27.2
 
 - Make HTTPS optional #192
+- Fix parsing .env file
 
 ## v0.27.1